Get rid of namespacing in specs.
This is a bad practice, since it can provide false positives in the test
suite due to the tests being scoped to the BSON module and not to the
global namespace.

Clients that use BSON will not be scoped to the BSON module, they will
be in their own namespace (like Mongoid, MongoMapper, etc), or in the
global scope. Out test suite should run in the same.